Remove vendored tokio-tungstenite crate
