+++ nautilus_core/Cargo.toml
@@ -12,7 +12,6 @@ members = [
     "infrastructure",
     "model",
     "network",
-    "network/tokio-tungstenite",
     "persistence",
     "pyo3",
     "cli"
