+++ nautilus_core/network/Cargo.toml
@@ -26,7 +26,10 @@ http = "1.1.0"
 hyper = "1.4.1"
 nonzero_ext = "0.3.0"
 reqwest = "0.12.7"
-tokio-tungstenite = { path = "./tokio-tungstenite", features = ["rustls-tls-native-roots"] }
+rustls = "0.23.12"
+tokio-rustls = "0.26.0"
+tokio-tungstenite = { version = "0.23.1", features = ["rustls-tls-native-roots"] }
+tungstenite = "0.24.0"
 
 [dev-dependencies]
 criterion = { workspace = true }
