--- nautilus_core/network/tokio-tungstenite/Cargo.toml
@@ -1,66 +1,0 @@
-[package]
-name = "tokio-tungstenite"
-description = "Tokio binding for Tungstenite, the Lightweight stream-based WebSocket implementation"
-categories = ["web-programming::websocket", "network-programming", "asynchronous", "concurrency"]
-keywords = ["websocket", "io", "web"]
-authors = ["Daniel Abramov <dabramov@snapview.de>", "Alexey Galakhov <agalakhov@snapview.de>"]
-license = "MIT"
-homepage = "https://github.com/snapview/tokio-tungstenite"
-documentation = "https://docs.rs/tokio-tungstenite/0.19.0"
-repository = "https://github.com/snapview/tokio-tungstenite"
-version = "0.23.1"
-edition = "2018"
-rust-version = "1.63"
-include = ["examples/**/*", "src/**/*", "LICENSE", "README.md", "CHANGELOG.md"]
-
-[package.metadata.docs.rs]
-features = ["native-tls", "__rustls-tls"]
-
-[features]
-default = ["connect", "handshake"]
-connect = ["stream", "tokio/net", "handshake"]
-handshake = ["tungstenite/handshake"]
-native-tls = ["native-tls-crate", "tokio-native-tls", "stream", "tungstenite/native-tls", "handshake"]
-native-tls-vendored = ["native-tls", "native-tls-crate/vendored", "tungstenite/native-tls-vendored"]
-rustls-tls-native-roots = ["__rustls-tls", "rustls-native-certs"]
-rustls-tls-webpki-roots = ["__rustls-tls", "webpki-roots"]
-__rustls-tls = ["rustls", "rustls-pki-types", "tokio-rustls", "stream", "tungstenite/__rustls-tls", "handshake"]
-stream = []
-
-[dependencies]
-log = "0.4.22"
-futures-util = { version = "0.3.30", default-features = false, features = ["sink", "std"] }
-tokio = { workspace = true }
-
-[dependencies.tungstenite]
-version = "0.24.0"
-default-features = false
-
-[dependencies.native-tls-crate]
-optional = true
-package = "native-tls"
-version = "0.2.11"
-
-[dependencies.rustls]
-optional = true
-version = "0.22.0"
-
-[dependencies.rustls-pki-types]
-optional = true
-version = "1.0"
-
-[dependencies.rustls-native-certs]
-optional = true
-version = "0.7.0"
-
-[dependencies.tokio-native-tls]
-optional = true
-version = "0.3.1"
-
-[dependencies.tokio-rustls]
-optional = true
-version = "0.25.0"
-
-[dependencies.webpki-roots]
-optional = true
-version = "0.26.1"
